+++ Userland/Libraries/LibJS/Runtime/PromiseResolvingElementFunctions.h
@@ -35,27 +35,43 @@ struct PromiseValueList final : public Cell {
     MarkedValueList values;
 };
 
+class PromiseResolvingElementFunction : public NativeFunction {
+    JS_OBJECT(PromiseResolvingFunction, NativeFunction);
+
+public:
+    virtual void initialize(GlobalObject&) override;
+    virtual ~PromiseResolvingElementFunction() override = default;
+
+    virtual Value call() override;
+
+protected:
+    explicit PromiseResolvingElementFunction(size_t, PromiseValueList&, PromiseCapability, RemainingElements&, Object& prototype);
+
+    virtual Value resolve_element() = 0;
+
+    size_t m_index { 0 };
+    PromiseValueList& m_values;
+    PromiseCapability m_capability;
+    RemainingElements& m_remaining_elements;
+
+private:
+    virtual void visit_edges(Visitor&) override;
+
+    bool m_already_called { false };
+};
+
 // 27.2.4.1.3 Promise.all Resolve Element Functions, https://tc39.es/ecma262/#sec-promise.all-resolve-element-functions
-class PromiseAllResolveElementFunction final : public NativeFunction {
+class PromiseAllResolveElementFunction final : public PromiseResolvingElementFunction {
     JS_OBJECT(PromiseResolvingFunction, NativeFunction);
 
 public:
     static PromiseAllResolveElementFunction* create(GlobalObject&, size_t, PromiseValueList&, PromiseCapability, RemainingElements&);
 
     explicit PromiseAllResolveElementFunction(size_t, PromiseValueList&, PromiseCapability, RemainingElements&, Object& prototype);
-    virtual void initialize(GlobalObject&) override;
     virtual ~PromiseAllResolveElementFunction() override = default;
 
-    virtual Value call() override;
-
 private:
-    virtual void visit_edges(Visitor&) override;
-
-    size_t m_index { 0 };
-    PromiseValueList& m_values;
-    PromiseCapability m_capability;
-    RemainingElements& m_remaining_elements;
-    bool m_already_called { false };
+    virtual Value resolve_element() override;
 };
 
 }
